requirements.txt
flask
python-docx
reportlab
werkzeug
